Figuring out a good way to detect objects
Don't want to check it every frame, so I decided to use an ienumerator
and run it every half a second, do a overlapsphere check to see which
colliders are in the area.

Also found out that having the patrol state as the first state isn't
going to work because it doesn't go through the enter function, so I
made an idle state to transition to patrol state for now. I needed to
create it anyway.